refactoring

+++ src/main/kotlin/me/aberrantfox/aegeus/listeners/InviteListener.kt
@@ -10,9 +10,14 @@ class InviteListener(val config: Configuration) : ListenerAdapter() {
         if(event.author.isBot) return
 
         if(event.message.rawContent.matches(Regex("(.|\n)*(https://discord.gg/)+(.|\n)*"))) {
+
+            if(event.message.rawContent.contains('@'))
+
             event.message.delete().queue()
             event.guild.textChannels.findLast { it.id == config.leaveChannel }
-                    ?.sendMessage("Deleted message: ${event.message.rawContent} by ${event.author.asMention}")
+                    ?.sendMessage("Deleted message: ${event.message.rawContent} " +
+                            "by ${event.author.asMention} " +
+                            "in ${event.channel.asMention}")
                     ?.queue()
         }
     }
